Modified classes
I modified the classes of the first question and answer to test if my JS code works at least for the first question. It doesn't. As soon as I have resolved this I will change the name of the classes of the other questions and answers and adapt the CSS file.